### 2024-02-19 #6374 @editorialbot
[PRE REVIEW]: Pybehave: a hardware-agnostic, Python-based framework for developing behavioral neuroscience experiments
**Submitting author:** @theonlydvr (<a href="http://orcid.org/0000-0002-1428-0723">Evan Dastin-van Rijn</a>)
**Repository:** https://github.com/tne-lab/py-behav-box-v2
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.16
