### 2024-02-26 #6374 @theonlydvr
For past JOSS papers/authors I think the following could work:

Grant-Giesbrecht: Hardware-Control: Instrument control and automation package
vsoch: The Experiment Factory: Reproducible Experiment Containers
ixjlyons: AxoPy: A Python Library for Implementing Human-Computer Interface Experiments
