### 2024-03-11 #6374 @sappelhoff
👋 @niksirbi & @sfmig & @tuliofalmeida & @vigji, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I will also try to send you a more formal invitation via email.
