### 2024-03-13 #6374 @sappelhoff
👋 @peircej & @hoechenberger & @isolver & @RebeccaHirst, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
