### 2024-03-13 #6374 @theonlydvr
The authors of [pycontrol](https://github.com/pyControl/code) could also be a good choice if you're considering people who haven't worked with JOSS before.
