### 2024-03-18 #6374 @sappelhoff
👋 @ThomasAkam & @alustig3 & @neuromantic99 & @AtMostafa, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
