### 2024-03-20 c961eca
Fix escaping of quote characters in source settings
