### 2024-03-20 7dba980
Added documentation for OSCAR
