+++ docs/tutorials/video_sync.md
@@ -16,9 +16,8 @@ interact with Video components expect additional metadata.
 
 The general purpose VideoSource is likely the go-to option for most users when synchronizing video. This source provides 
 a general framework for displaying and serializing video feeds with specific implementations for various feed types. Currently,
-this source only supports webcams and ImagingSource cameras but more will likely be added in the future. Webcams use the 
-camera index as their address while ImagingSource cameras have a more complex format that is described in detail in their 
-documentation. If you are interested in adding video feeds from non-implemented systems, simply override the abstract VideoProvider
+this source only supports webcams and Imaging Source cameras but more will likely be added in the future. 
+If you are interested in adding video feeds from non-implemented systems, simply override the abstract VideoProvider
 class and take the existing options as examples.
 
 The VideoSource can support an arbitrary number of camera feeds which will be displayed in a grid in a companion GUI. To set up the source,
@@ -31,27 +30,79 @@ AddressFile: which row 'row' and column 'col' in the grid the video should be pl
 additional columns 'col_span' the video will take up in the grid (0 for a single cell), and the desired frame rate 'fr'. 
 The video will be saved at whatever framerate is specified even if the underlying video provider is generating frames faster 
 or slower (frames will be skipped or duplicated). This is done to ensure the video will exactly match the time elapsed between 
-when it was started and stopped. An example AddressFile line for a webcam is shown below:
+when it was started and stopped. If you want to use a VideoProvider other than the webcam, you will also need to add a 'vid_type'
+key-value pair. Currently, the following provider designations are supported:
+
+    webcam: WebcamProvider (default)
+    imaging_source: ImagingSourceProvider
+    
+
+### Identifying the video feed address
+
+VideoProviders require different formats for the addresses of the corresponding video feeds. 
+
+#### Webcams
+
+We use OpenCV to implement video capture with webcams. OpenCV assigns each camera connected to the computer an index with 
+the system default camera having index '0'. This index should be used as the address when declaring the pybehave component.
+OpenCV has no consistent system for assigning indices to cameras, so we recommend trying every index beginning from '0' if 
+you have multiple cameras connected to the system and want to identify the correct one.
+
+An example AddressFile line for a webcam is shown below:
 
     addresses.add_component("cam", "Video", "vs", "0", None, {"fr":30,"row":0,"col":0, "col_span":0,"row_span":0})
 
+#### Imaging Source
+
+[The Imaging Source](https://www.theimagingsource.com/en-us/) manufactures high-end industrial cameras that have been used
+in some animal research applications. Using their provided [API](https://github.com/TheImagingSource/IC-Imaging-Control-Samples/tree/master/Python), 
+we've integrated acquisition from Imaging Source cameras into the pybehave VideoSource framework. When using these cameras,
+the address is full serial number for the corresponding camera. 
+
+An example AddressFile line for an Imagine Source camera is shown below:
+
+    addresses.add_component("cam", "Video", "vs", "DFK Z30GP031 41910044", None, {"fr":30,"row":0,"col":0, "col_span":0,"row_span":0,"vid_type":"imaging_source"})
+
+
 ## HikVisionSource
 
-One option that the TNE lab has had a lot of success with is using traditional security system (DVR) hardware to manage out
+One option that the TNEL has had a lot of success with is using traditional security system (DVR) hardware to manage our
 video recordings. These systems are very reliable, support many feeds, and are far less expensive than traditional researcher-facing
 tools. In particular, we use HikVision DVRs because they have a well documented and robust API for controlling recordings
-over the local network. 
+over the local network.
 
-To set up the HikVisionSource, you will need to provide the IP of the device (this can be found from their companion software)
-and the username and password for the DVR.
+### Setup
+
[…]
