### 2024-04-02 tne-lab/py-behav-box-v2#68 @tuliofalmeida
Review Joss Paper Tulio 
Hello, first of all it's a pleasure to [review](https://github.com/openjournals/joss-reviews/issues/6515) your work!

So, so far I'm checking it alongside with experimentalist friend that doesn't know how to code but would like to use the library.
I would like to know who is the target of your library? Because I want to test it with this friend that will need it to plan the experiments.

In the first testing session, we did today was frustrating to him (basic python level) that you can't do everything trough the interface. We tried to follow the tutorial to add the webcam, and he got lost in creating the AdresseFile, is not clear in the tutorial how you switch to the GUI to create the file.

If you are aiming advance people in python already adapted to do this setups, I'll change my perspective.

Thanks
