### 2024-04-02 tne-lab/py-behav-box-v2#68 @theonlydvr
I'd say my intended target was generally experimentalists who had some code familiarity (so some basic Python experience at least). However, I don't think that group is necessarily that different from experimentalists who have to use existing behavioral systems (Med Associates, ABET, etc) because they often have their own custom language-like structures which have a large learning curve. Generally, I'd like to make the system workable for those users at least as far as being able to relatively easily setup an existing task on a new rig.

So given what you're describing, it sounds like the frustrating portion was that the AddressFile and Protocol creation had to happen in a text editor rather than through a GUI tool? I think it could be reasonable to add an interface to create new AddressFiles or protocols for an input task. Do you think that would address the need?

I'm very happy to add more features along these lines to streamline the process for this sort of user.
