### 2024-04-02 5aee668
Added protocol support to validator
