+++ mkdocs.yml
@@ -17,6 +17,5 @@ nav:
     - Logging events: events.md
     - Developing components: components.md
     - Connecting to hardware: sources.md
-    - TaskSequences: task_sequences.md
 
 copyright: Copyright &copy; 2021 - 2023 Translational Neuroengineering Laboratory
