### 2024-04-03 tne-lab/py-behav-box-v2#68 @tuliofalmeida
Perfect. First was not that clear that was necessary to do this separated in text file. But, this solution that you propose is really good and maybe can overcome one future problem with paths and files allocated in wrong folders. So you agree that we can follow this line of thinking for the testing ?
