### 2024-04-03 tne-lab/py-behav-box-v2#68 @theonlydvr
Definitely a good line of thinking for testing, the combination of your expertise with an experimentalist user will be very helpful. On the earlier point, I'll make sure to clarify the existing documentation and workshop the GUI option.
