+++ pybehave/Workstation/AddressFileCreationDialog.py
@@ -1,0 +1,317 @@
+from __future__ import annotations
+
+import bisect
+import copy
+import importlib
+import os
+import pkgutil
+from typing import TYPE_CHECKING
+
+from PyQt5 import QtCore
+from PyQt5.QtCore import QSize
+from PyQt5.QtWidgets import QDialog, QDialogButtonBox, QVBoxLayout, QTableWidget, QPushButton, QHBoxLayout, \
+    QTableWidgetItem, QComboBox, QMessageBox, QFileDialog
+
+if TYPE_CHECKING:
+    from pybehave.Workstation.WorkstationGUI import WorkstationGUI
+
+import pybehave.Components
+
+
+class ComboBox(QComboBox):
+    new_signal = QtCore.pyqtSignal(str, str)
+
+    def __init__(self, parent=None):
+        super(ComboBox, self).__init__(parent)
+        self.lastSelected = None
+        self.activated[str].connect(self.onActivated)
+
+    def onActivated(self, text):
+        self.new_signal.emit(self.lastSelected, text)
+        self.lastSelected = text
+
+
+class NoBlankSpaceAtBottomEvenlySplitTableView(QTableWidget):
+    def sizeHint(self):
+        w = self.verticalHeader().width() + 4  # +4 seems to be needed
+        for i in range(self.columnCount()):
+            w += self.columnWidth(i)
+        height = 0
+        if self.horizontalHeader().isVisible():
+            height += self.horizontalHeader().height()
+        height += self.verticalHeader().length() + self.frameWidth() * 2
+        return QSize(w, height)
+
+
+def get_all_subclasses(cls):
+    all_subclasses = []
+
+    for subclass in cls.__subclasses__():
+        all_subclasses.append(subclass)
+        all_subclasses.extend(get_all_subclasses(subclass))
+
+    return all_subclasses
+
+
+class AddressFileCreationDialog(QDialog):
+    def __init__(self, wsg: WorkstationGUI, task: str):
+        super().__init__()
+
+        self.wsg = wsg
+
+        task_module = importlib.import_module("Local.Tasks." + task)
+        self.task = getattr(task_module, task)
+
+        self.components = self.task.get_components()
+        self.available_components = {}
+        for key in self.components:
+            if len(self.components[key]) > 1:
+                self.available_components[key] = [str(i) for i in range(len(self.components[key]))]
+            else:
+                self.available_components[key] = None
+
+        for f in pkgutil.iter_modules(pybehave.Components.__path__):
+            if f.name[0].isupper():
+                importlib.import_module("pybehave.Components." + f.name, f.name)
+
+        self.setWindowTitle(task + " AddressFile")
+        self.setMinimumSize(500, 700)
+        control = QDialogButtonBox.Save | QDialogButtonBox.Cancel
+
+        button_layout = QHBoxLayout()
+        self.control_buttons = QDialogButtonBox(control)
+        self.control_buttons.accepted.connect(self.save)
+        self.control_buttons.rejected.connect(self.reject)
+        self.add_button = QPushButton("+")
+        self.add_button.clicked.connect(self.add_row)
+        button_layout.addWidget(self.add_button, alignment=QtCore.Qt.AlignLeft)
+        button_layout.addWidget(self.control_buttons, alignment=QtCore.Qt.AlignRight)
+
+        self.layout = QVBoxLayout()
+        self.table = QTableWidget()
+
+        self.table.setColumnCount(6)
+        self.table.setHorizontalHeaderLabels(["Component", "Type", "Source", "Address", "Index", "Metadata"])
+        self.table.verticalHeader().setVisible(False)
+
+        self.layout.addWidget(self.table)
+        self.layout.addLayout(button_layout)
[…]
