### 2024-04-07 39a671f
Added protocol creation GUI and task selection dialog
