### 2024-04-11 tne-lab/py-behav-box-v2#68 @theonlydvr
I just pushed some code to the dev branch that provides an option for creating AddressFiles and Protocols using a GUI (they can be accessed from the file menu). I'm still working on adding documentation for the feature but let me know if it lines up with what you thought might be helpful.
