+++ mkdocs.yml
@@ -18,4 +18,4 @@ nav:
     - Developing components: components.md
     - Connecting to hardware: sources.md
 
-copyright: Copyright &copy; 2021 - 2023 Translational Neuroengineering Laboratory
+copyright: Copyright &copy; 2021 - 2024 Translational Neuroengineering Laboratory
