### 2024-04-18 e5268d0
Allowed None for list_index with singular components
