### 2024-04-19 7334443
Added missing extra handling and documentation links
