+++ pybehave/Workstation/SettingsDialog.py
@@ -2,9 +2,11 @@ from __future__ import annotations
 
 import multiprocessing
 import os
+import webbrowser
 from typing import TYPE_CHECKING
 
 from pybehave.Events.PybEvents import AddSourceEvent, RemoveSourceEvent
+from pybehave.Utilities.Exceptions import MissingExtraError
 from pybehave.Utilities.find_closing_paren import find_closing_paren
 import pybehave.Sources
 
@@ -83,6 +85,7 @@ class SettingsDialog(QDialog):
         self.layout.addWidget(source_box)
         self.layout.addWidget(self.control_buttons)
         self.setLayout(self.layout)
+        self.error_dialog = None
 
     def update_refresh_check(self):
         self.workstation.refresh_gui = self.refresh_gui.isChecked()
@@ -154,6 +157,7 @@ class SettingsDialog(QDialog):
 
 
 class AddSourceDialog(QDialog):
+
     def __init__(self, sd: SettingsDialog):
         super(AddSourceDialog, self).__init__()
         self.sd = sd
@@ -193,18 +197,35 @@ class AddSourceDialog(QDialog):
         self.setLayout(self.layout)
         self.params = []
 
+    def event(self, event):
+        if event.type() == QEvent.EnterWhatsThisMode:
+            QWhatsThis.leaveWhatsThisMode()
+            webbrowser.open('https://py-behav-box-v2.readthedocs.io/en/latest/sources/#included-sources')
+            return True
+        else:
+            return super().event(event)
+
     def set_params(self) -> None:
-        if self.source.currentText() in self.sources:
-            source_type = getattr(importlib.import_module("pybehave.Sources." + self.source.currentText()), self.source.currentText())
-        else:
-            source_type = getattr(importlib.import_module("Local.Sources." + self.source.currentText()),
-                                  self.source.currentText())
-        all_params = inspect.getfullargspec(source_type.__init__)
-        if len(all_params.args) > 1:
-            self.spd = SourceParametersDialog(self, all_params)
-            self.spd.show()
-        else:
-            self.accept()
+        try:
+            if self.source.currentText() in self.sources:
+                source_type = getattr(importlib.import_module("pybehave.Sources." + self.source.currentText()), self.source.currentText())
+            else:
+                source_type = getattr(importlib.import_module("Local.Sources." + self.source.currentText()),
+                                      self.source.currentText())
+            all_params = inspect.getfullargspec(source_type.__init__)
+            if len(all_params.args) > 1:
+                self.spd = SourceParametersDialog(self, all_params)
+                self.spd.show()
+            else:
+                self.accept()
+        except MissingExtraError as e:
+            self.sd.error_dialog = QMessageBox(self)
+            self.sd.error_dialog.setWindowTitle('Missing extra')
+            self.sd.error_dialog.setText(f'{self.source.currentText()} requires the \'{e.extra}\' extra. This extra can be installed by running \'pip install pybehave[{e.extra}]\' (or \'pip install .[{e.extra}]\' if you\'re using a local installation) in your pybehave virtual environment.')
+            self.sd.error_dialog.setStandardButtons(QMessageBox.Ok)
+            self.sd.error_dialog.setIcon(QMessageBox.Critical)
+            self.sd.error_dialog.show()
+            self.close()
 
     def escape(self, s: str) -> str:
         s = s.replace('\'', '\\\\\\\'')
@@ -276,6 +297,14 @@ class SourceParametersDialog(QDialog):
             self.params.append(param)
         self.layout.addWidget(self.control_buttons)
         self.setLayout(self.layout)
+
+    def event(self, event):
+        if event.type() == QEvent.EnterWhatsThisMode:
[…]
