+++ pybehave/Workstation/ConfigurationDialog.py
@@ -5,6 +5,7 @@ from typing import TYPE_CHECKING
 
 from pybehave.Events.PybEvents import AddLoggerEvent, RemoveLoggerEvent
 import pybehave.Events
+from pybehave.Utilities.Exceptions import MissingExtraError
 
 if TYPE_CHECKING:
     from pybehave.Workstation.ChamberWidget import ChamberWidget
@@ -91,6 +92,7 @@ class ConfigurationDialog(QDialog):
 
         self.layout.addWidget(self.control_buttons)
         self.setLayout(self.layout)
+        self.error_dialog = None
     
     def accept(self) -> None:
         self.cw.prompt = self.prompt.text()  # Update the prompt from the configuration
@@ -171,13 +173,22 @@ class AddExtrasDialog(QDialog):
         super(AddExtrasDialog, self).accept()
 
     def set_params(self) -> None:
-        logger_type = getattr(importlib.import_module("pybehave.Events." + self.extra.currentText()), self.extra.currentText())
-        all_params = inspect.getfullargspec(logger_type.__init__)
-        if len(all_params.args) > 1:
-            self.epd = ExtrasParametersDialog(self, all_params)
-            self.epd.show()
-        else:
-            self.accept()
+        try:
+            logger_type = getattr(importlib.import_module("pybehave.Events." + self.extra.currentText()), self.extra.currentText())
+            all_params = inspect.getfullargspec(logger_type.__init__)
+            if len(all_params.args) > 1:
+                self.epd = ExtrasParametersDialog(self, all_params)
+                self.epd.show()
+            else:
+                self.accept()
+        except MissingExtraError as e:
+            self.cd.error_dialog = QMessageBox(self)
+            self.cd.error_dialog.setWindowTitle('Missing extra')
+            self.cd.error_dialog.setText(f'{self.extra.currentText()} requires the \'{e.extra}\' extra. This extra can be installed by running \'pip install pybehave[{e.extra}]\' (or \'pip install .[{e.extra}]\' if you\'re using a local installation) in your pybehave virtual environment.')
+            self.cd.error_dialog.setStandardButtons(QMessageBox.Ok)
+            self.cd.error_dialog.setIcon(QMessageBox.Critical)
+            self.cd.error_dialog.show()
+            self.close()
 
 
 class ExtrasParametersDialog(QDialog):
