### 2024-04-19 182dcca
Added documentation links to events/widgets
