### 2024-04-23 cc15bc7
Added information about cross-platform support
