### 2024-04-24 38b6a2c
Fixed closing behavior for OSCAR and editing for AddressFile GUI
