### 2024-04-26 5820c64
Added more explicit documentation on using pybehave without hardware
