+++ docs/guis.md
@@ -5,6 +5,8 @@
 Pybehave provides a framework for customizable monitoring and representation of task behavior using task-specific graphical
 user interfaces (GUIs). GUIs are written using the [pygame]() library with individual objects in the GUI wrapped as *Element*
 classes. GUIs can be used both for visually presenting task data or allow experimenters to interact with the task workflow.
+This second capacity is especially useful because GUIs can be used to test and develop tasks without a connection to external
+hardware.
 
 ## Elements
 
