### 2024-04-29 #6515 @tuliofalmeida
> @alustig3 @tuliofalmeida could I get a brief status update of your reviews, please?
>
> @theonlydvr could you provide an update from your side as well, please? Are there any questions about the review process, or anything we should discuss/clarify?

I'm finishing! The authors did updates in the package (very quickly) but I didn't had time to check it yet. I'm planning to finish this week.
