### 2024-04-30 f15460a
Fixed bug related to qasync failing with non-awaited tasks
https://github.com/CabbageDevelopment/qasync/issues/121
