### 2024-04-30 tne-lab/py-behav-box-v2#68 @theonlydvr
A few changes were made to the VideoSource on a recent commit that will avoid some bugs you would likely have come across.
