### 2024-04-30 6d9cb94
make chamber count a QSpinbox
