### 2024-04-30 5201e59
add permalinks to docs
this makes it easier to share links to specific sections within the docs
