### 2024-05-01 6365453
Set explicit minimum for QSpinBox
