### 2024-05-01 75ad0ab
Implemented in-session constant adjustment with the SubjectConfigWidget
