### 2024-05-13 tne-lab/py-behav-box-v2#68 @tuliofalmeida
Hey, I tested the new feature I think that will be way more easy to non-coders. I'll finish the review, but I'll suggest to my lab mates to use it in the next experiments.
