### 2024-05-13 #6515 @tuliofalmeida
Hey @sappelhoff I finished my part of the review. The idea was to test it with a lab friend that do experiments but doesn't code. @theonlydvr did a good job, my main point is to improve the documentation and tutorial for people that are not comfortable with coding and he's doing it.
