### 2024-05-13 #6515 @sappelhoff
Thanks for the note @tuliofalmeida!

@alustig3 @tuliofalmeida, I can see you both checked your reviewer lists fully. I'd appreciate it, **if you could briefly make a statement whether you would recommend this paper for publication, or whether there are items that need to be dealt with before.** For example, in @alustig3's reviewer list, I see some linked PRs/issues that have not yet been merged/closed, and @tuliofalmeida mentions:

> ... and he's doing it.

... which suggest to me, that there are still some things to be done.

In any way, thanks already for the work until here, it seems like we are nearing the completion of this review process.
