### 2024-05-21 b0b66a9
Made labels more explicit for SubjectConfigWidget
