+++ pybehave/Events/SubjectConfigWidget.py
@@ -24,8 +24,7 @@ class SubjectConfigWidget(EventWidget):
         self.layout.addWidget(self.widget)
         self.setLayout(self.layout)
         self.config_layout = QVBoxLayout(self.widget)
-        self.buttons_layout = QVBoxLayout(self.widget)
-        self.controls_layout = QHBoxLayout(self.widget)
+        self.controls_layout = QVBoxLayout(self.widget)
         self.specific_layout = QVBoxLayout(self.widget)
         self.top_layout = QHBoxLayout(self.widget)
         self.protocol_specific = QCheckBox("Protocol Specific")
@@ -35,24 +34,20 @@ class SubjectConfigWidget(EventWidget):
         self.specific_layout.addWidget(self.protocol_specific)
         self.specific_layout.addWidget(self.address_specific)
         self.remove_button = QPushButton()
-        self.remove_button.setText("−")
-        self.remove_button.setFixedWidth(30)
+        self.remove_button.setText("Remove Constant")
         self.remove_button.setDisabled(True)
         self.remove_button.clicked.connect(self.remove_constant)
         self.add_button = QPushButton()
-        self.add_button.setText("+")
-        self.add_button.setFixedWidth(30)
+        self.add_button.setText("Add Constant")
         self.add_button.clicked.connect(self.add_constant)
-        self.controls_layout.addWidget(self.remove_button)
         self.controls_layout.addWidget(self.add_button)
-        self.buttons_layout.addLayout(self.controls_layout)
+        self.controls_layout.addWidget(self.remove_button)
         self.save_button = QPushButton()
-        self.save_button.setText("↓")
-        self.save_button.setFixedWidth(30)
+        self.save_button.setText("Save Constants")
         self.save_button.clicked.connect(self.save_values)
-        self.buttons_layout.addWidget(self.save_button, alignment=Qt.AlignHCenter)
+        self.controls_layout.addWidget(self.save_button)
         self.top_layout.addLayout(self.specific_layout)
-        self.top_layout.addLayout(self.buttons_layout)
+        self.top_layout.addLayout(self.controls_layout)
         self.config_layout.addLayout(self.top_layout)
         constants_name_box = QGroupBox('Name')
         constants_name_layout = QVBoxLayout(self.widget)
