### 2024-05-21 e3aff9b
Added error handling for GUI errors
