+++ pybehave/Workstation/Workstation.py
@@ -4,6 +4,7 @@ import multiprocessing
 import sys
 import threading
 import time
+import traceback
 from multiprocessing.dummy.connection import Connection
 from typing import List
 
@@ -213,102 +214,110 @@ class Workstation:
             for ready in multiprocessing.connection.wait(conns):
                 events = self.decoder.decode(ready.recv_bytes())
                 for event in events:
-                    if isinstance(event, PybEvents.AddTaskEvent):
-                        module = importlib.import_module("Local.GUIs." + event.task_name + "GUI")
-                        module = importlib.reload(module)
-                        gui = getattr(module, event.task_name + "GUI")
-                        # Position the GUI in pygame
-                        col = event.chamber % self.n_col
-                        row = math.floor(event.chamber / self.n_col)
-                        # Create the GUI
-                        self.guis[event.chamber] = gui(event, self.task_gui.subsurface(col * self.w, row * self.h, self.w, self.h), self)
-                    elif isinstance(event, PybEvents.TaskEvent):
-                        if event.chamber in self.guis:
-                            for widget in self.wsg.chambers[event.chamber].widgets:
-                                if isinstance(widget, EventWidget):
-                                    widget.emitter.emit(event)
-                            self.guis[event.chamber].handle_event(event)
+                    try:
+                        if isinstance(event, PybEvents.AddTaskEvent):
+                            module = importlib.import_module("Local.GUIs." + event.task_name + "GUI")
+                            module = importlib.reload(module)
+                            gui = getattr(module, event.task_name + "GUI")
+                            # Position the GUI in pygame
                             col = event.chamber % self.n_col
                             row = math.floor(event.chamber / self.n_col)
-                            rect = pygame.Rect((col * self.w, row * self.h, self.w, self.h))
-                            if isinstance(event, PybEvents.InitEvent) or isinstance(event, PybEvents.StartEvent):
-                                if "sub_task" in event.metadata:
-                                    self.guis[event.chamber].switch_sub_gui(event)
-                                self.guis[event.chamber].complete = False
-                                self.guis[event.chamber].draw()
-                                self.gui_updates.append(rect)
-                            elif isinstance(event, PybEvents.OutputFileChangedEvent):
-                                self.guis[event.chamber].subject_name.text = event.subject
-                                pygame.draw.rect(self.guis[event.chamber].task_gui, Colors.darkgray, self.guis[event.chamber].subject_name.rect, 0)
-                                self.guis[event.chamber].subject_name.draw()
-                                self.gui_updates.append(self.guis[event.chamber].subject_name.rect.move(col * self.w, row * self.h))
-                            elif isinstance(event, PybEvents.TaskCompleteEvent):
-                                if not isinstance(self.guis[event.chamber], SequenceGUI) or "sequence_complete" in event.metadata:
-                                    self.guis[event.chamber].complete = True
+                            # Create the GUI
+                            self.guis[event.chamber] = gui(event, self.task_gui.subsurface(col * self.w, row * self.h, self.w, self.h), self)
+                        elif isinstance(event, PybEvents.TaskEvent):
+                            if event.chamber in self.guis:
+                                for widget in self.wsg.chambers[event.chamber].widgets:
+                                    if isinstance(widget, EventWidget):
+                                        widget.emitter.emit(event)
+                                self.guis[event.chamber].handle_event(event)
+                                col = event.chamber % self.n_col
+                                row = math.floor(event.chamber / self.n_col)
+                                rect = pygame.Rect((col * self.w, row * self.h, self.w, self.h))
+                                if isinstance(event, PybEvents.InitEvent) or isinstance(event, PybEvents.StartEvent):
+                                    if "sub_task" in event.metadata:
+                                        self.guis[event.chamber].switch_sub_gui(event)
[…]
