### 2024-05-21 2d25b6c
Added context menu items for removing protocols/addressfiles
