### 2024-05-22 #6515 @sappelhoff
@theonlydvr I have generated the final steps that you and I need to go through, see: https://github.com/openjournals/joss-reviews/issues/6515#issuecomment-2124051446

Could you please take care of your items (upper list), and let me know once you are finished?
