### 2024-05-22 #6515 @theonlydvr
Authors and affiliations have been checked.
Version number is 1.0.19 on both the Github and pypi releases.
DOI was created with Zenodo and all information is consistent with the JOSS submission: https://doi.org/10.5281/zenodo.11244351
