### 2024-06-02 #6515 @sappelhoff
@theonlydvr I have completed my final checks of this submission. Before I can instruct the bot to "recommend-accept", I would like to clarify the following two points:

- [ ] The software version in the archive (as can be seen in `pyproject.toml`) is `1.0.18`, and not `1.0.19`, as I expected. Did something go wrong during the release / archiving?

- [ ] I see that very few of the authors on your author list have an associated ORCID. However, a quick check shows me that some of these authors do in fact have an ORCID (for example https://orcid.org/0000-0002-1669-8752, if I am not mistaken). Could you please add all ORCIDs that are available, unless the respective authors explicitly don't want to list them?
