### 2024-06-03 #6515 @theonlydvr
So it looks like the release grabbed the commit that preceded me updating the version number in the toml file but has all the correct code and content. It should be correct otherwise but I can make a new release and send the information over if necessary. Most of the authors are technical staff and trainees who helped with development and testing so they don't have ORCID IDs yet. I did add two new ORCID IDs for the ones that do though.

I have also now added myself to the reviewer database 🙂
