### 2024-06-05 #6515 @Kevin-Mattheus-Moerman
@theonlydvr as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. I only have the below point that requires your attention:

- [ ] Please spell out USA as United States of America in your affiliations.
