### 2024-06-06 642e7f5
Create CITATION.cff
