### 2024-06-06 #6515 @theonlydvr
Awesome! Everything looks good except it seems like my last name "Dastin-van Rijn" autogenerated to only have "Rijn" as the last name in the citation metadata. I've corrected this in the CITATION.cff in the original repository but I assume this has no effect on the actual merged paper. Apologies for the space in the last name...
