### 2024-06-06 #6515 @Kevin-Mattheus-Moerman
@sappelhoff @theonlydvr yes we can take it from here. I'll ping @openjournals/dev to see if they can offer some guidance on this one.
