### 2024-06-06 #6515 @xuanxu
The parts of the name can be described with more precision in the paper's metadata:
https://joss.readthedocs.io/en/latest/paper.html#names

Updating the paper.md file and then reaccepting should update the published strings.
