+++ paper/paper.md
@@ -5,7 +5,8 @@ tags:
   - Animal behavior
   - Operant tasks
 authors:
-  - name: Evan M. Dastin-van Rijn
+  - given-names: Evan M.
+    surname: Dastin-van Rijn
     orcid: 0000-0002-1428-0723
     affiliation: 1
     corresponding: true
